Create MediaPlayer package (#543)
* initial move

* App with media player booting

* Working media player into package

* Fixed storybook

* Rename mediaplayer -> media-player

* Fix tests

* Rename mediaplayer folder to media-player

* Create tab bar wrapper

* Fix media player state thrashing

* Added readme

* Fix tests

* Fix tests
