+++ packages/apollos-ui-auth/package.json
@@ -53,7 +53,7 @@
   },
   "devDependencies": {
     "apollo-client": "2.5.1",
-    "jest": "23.6.0",
+    "jest": "24.8.0",
     "jest-react-native": "18.0.0",
     "prop-types": "15.6.1",
     "react": "^16.8.3",
