+++ packages/apollos-ui-htmlview/package.json
@@ -42,7 +42,7 @@
     "htmlparser2": "3.9.2"
   },
   "devDependencies": {
-    "jest": "23.6.0",
+    "jest": "24.8.0",
     "jest-react-native": "18.0.0",
     "prop-types": "15.6.1",
     "react": "16.8.3",
