+++ packages/apollos-ui-kit/package.json
@@ -57,7 +57,7 @@
     "rn-placeholder": "1.2.0"
   },
   "devDependencies": {
-    "jest": "23.6.0",
+    "jest": "24.8.0",
     "jest-react-native": "18.0.0",
     "prop-types": "15.6.1",
     "react": "16.8.3",
