+++ packages/apollos-ui-media-player/jest.setup.js
@@ -1,0 +1,36 @@
+jest.mock('@apollosproject/ui-analytics', () => ({
+  track: () => '',
+  AnalyticsProvider: ({ children }) => children,
+}));
+
+jest.mock('react-native-airplay-btn', () => ({
+  AirPlayButton: () => 'AirPlayButton',
+}));
+
+jest.mock('react-native-music-control', () => ({
+  enableBackgroundMode: jest.fn(),
+  enableControl: jest.fn(),
+  on: jest.fn(),
+  setNowPlaying: jest.fn(),
+}));
+
+jest.mock('Animated', () => {
+  const ActualAnimated = require.requireActual('Animated');
+  return {
+    ...ActualAnimated,
+    timing: (value, config) => ({
+      start: (callback) => {
+        value.setValue(config.toValue);
+        callback && callback();
+      },
+      stop: () => ({}),
+    }),
+    spring: (value, config) => ({
+      start: (callback) => {
+        value.setValue(config.toValue);
+        callback && callback();
+      },
+      stop: () => ({}),
+    }),
+  };
+});
