+++ packages/apollos-ui-media-player/src/MediaPlayer/MediaPlayerSpacer.js
@@ -1,33 +1,25 @@
 import React from 'react';
-import { SafeAreaView } from 'react-native';
 import { Query } from 'react-apollo';
-import gql from 'graphql-tag';
 import { get } from 'lodash';
 
 import { styled } from '@apollosproject/ui-kit';
-import { MINI_PLAYER_HEIGHT } from 'apolloschurchapp/src/ui/MediaPlayer';
+import { SafeAreaView } from 'react-navigation';
+import { MINI_PLAYER_HEIGHT } from './MiniControls';
 
 import MediaPlayerSafeLayout from './MediaPlayerSafeLayout';
-
-const mediaPlayerIsVisibleQuery = gql`
-  query {
-    mediaPlayer @client {
-      isVisible
-    }
-  }
-`;
+import { getMediaPlayerVisibility } from './queries';
 
 const MediaPlayerSafeLayoutWithSpacing = styled({
   paddingBottom: MINI_PLAYER_HEIGHT,
 })(MediaPlayerSafeLayout);
 
 const MediaPlayerSpacer = (props) => (
-  <Query query={mediaPlayerIsVisibleQuery}>
+  <Query query={getMediaPlayerVisibility}>
     {({ data = {} }) =>
       get(data, 'mediaPlayer.isVisible') ? (
         <MediaPlayerSafeLayoutWithSpacing {...props} />
       ) : (
-        <SafeAreaView {...props} />
+        <SafeAreaView forceInset={{ bottom: 'always' }} {...props} />
       )
     }
   </Query>
