+++ packages/apollos-ui-media-player/src/MediaPlayer/MiniControls.js
