+++ packages/apollos-ui-media-player/src/MediaPlayer/MusicControls.js
