+++ packages/apollos-ui-media-player/src/MediaPlayer/VideoWindow.js
