+++ packages/apollos-ui-media-player/src/MediaPlayer/queries.js
