+++ packages/apollos-ui-onboarding/package.json
@@ -29,6 +29,7 @@
     "modulePaths": [
       "<rootDir>/src/"
     ],
+    "modulePathIgnorePatterns": ["node_modules/@storybook/"],
     "setupFiles": [
       "./jest.setup.js"
     ]
@@ -54,7 +55,7 @@
   "devDependencies": {
     "@storybook/react-native": "^4.0.7",
     "apollo-client": "2.5.1",
-    "jest": "23.6.0",
+    "jest": "24.8.0",
     "jest-react-native": "18.0.0",
     "prop-types": "15.6.1",
     "react": "16.8.3",
