+++ packages/apollos-ui-passes/package.json
@@ -29,6 +29,7 @@
     "setupFiles": [
       "./jest.setup.js"
     ],
+    "modulePathIgnorePatterns": ["node_modules/@storybook/"],
     "modulePaths": [
       "<rootDir>/src/"
     ]
@@ -47,7 +48,7 @@
     "react-apollo": "2.5.5"
   },
   "devDependencies": {
-    "jest": "23.6.0",
+    "jest": "24.8.0",
     "jest-react-native": "18.0.0",
     "prop-types": "15.6.1",
     "react": "16.8.3",
