+++ packages/apolloschurchapp/metro.config.js
@@ -8,6 +8,9 @@
 const path = require('path');
 const blacklist = require('metro-config/src/defaults/blacklist');
 
+const flatten = (array) =>
+  array.reduce((accum, curr) => curr.concat(accum), []);
+
 const localDeps = [
   path.resolve('..', 'apollos-ui-kit'),
   path.resolve('..', 'apollos-ui-passes'),
@@ -16,17 +19,29 @@ const localDeps = [
   path.resolve('..', 'apollos-ui-analytics'),
   path.resolve('..', 'apollos-ui-storybook'),
   path.resolve('..', 'apollos-ui-onboarding'),
+  path.resolve('..', 'apollos-ui-media-player'),
   path.resolve('..', 'apollos-ui-scripture'),
 ];
 
+const sharedNativeModules = [
+  'react-native-linear-gradient',
+  'react-native-svg',
+  'react-native-music-control',
+  'react-native-video',
+  'react-native-video-controls',
+  'react-native-airplay-btn',
+];
 
 module.exports = {
   resolver: {
     extraNodeModules: {
       'react-native': path.resolve('.', 'node_modules/react-native'),
-      'react-native-linear-gradient': path.resolve(
-        '.',
-        'node_modules/react-native-linear-gradient'
+      ...sharedNativeModules.reduce(
+        (accum, curr) => ({
+          ...accum,
+          [curr]: path.resolve('.', `node_modules/${curr}`),
+        }),
+        {}
       ),
       'react-native-svg': path.resolve('.', 'node_modules/react-native-svg'),
     },
@@ -34,9 +49,17 @@ module.exports = {
       ...localDeps.map(
         (depPath) => new RegExp(`${depPath}/node_modules/react-native/.*`)
       ),
-      ...localDeps.map(
-        (depPath) =>
-          new RegExp(`${depPath}/node_modules/react-native-linear-gradient/.*`)
+      ...flatten(
+        localDeps.map((depPath) =>
+          sharedNativeModules.map(
+            (moduleName) =>
+              new RegExp(`${depPath}/node_modules/${moduleName}/.*`)
+          )
+        )
+      ),
+      ...sharedNativeModules.map(
+        (moduleName) =>
+          new RegExp(`${path.resolve(`../../node_modules/${moduleName}`)}/.*`)
       ),
       ...localDeps.map(
         (depPath) => new RegExp(`${depPath}/node_modules/react-native-svg/.*`)
@@ -49,12 +72,6 @@ module.exports = {
         `${path.resolve('.', 'node_modules')}/.*/node_modules/react-native/.*`
       ),
       new RegExp(
-        `${path.resolve(
-          '.',
-          'node_modules'
-        )}/react-native-iphone-x-helper/node_modules/react-native/.*`
-      ),
-      new RegExp(
         `${path.resolve('..', '..', 'node_modules', 'react-native')}/.*`
       ),
       new RegExp(
@@ -63,9 +80,6 @@ module.exports = {
           '..',
           'node_modules'
         )}/.*/node_modules/react-native/.*`
-      ),
-      new RegExp(
-        `${path.resolve('../../node_modules/react-native-linear-gradient')}/.*`
       ),
     ]),
   },
